requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
supabase>=2.0.0
anthropic>=0.25.0
playwright>=1.40.0
apscheduler>=3.10.0
